major overhaul of device class, carrying now the  project description
moved the ViennaFVM::project_description into the device class,
same variant approach as with the viennagrid::mesh.

the device now assigns the doping via the viennafvm mechanism,
also the epsr is automatically assigned as soon as a material is assigned
to a segment.

-> the problem implementation (laplace, poisson_dd-np) are now much simpler,
as they require the presence of, e.g., the permitivity in the project description.

-> this approach is now similar to ViennaSHE approach

all examples work