investigating SRH problem: suspecting ViennaMath/FVM problem
the problem is the two terms in the denominator (regardless if they
are on the left or right side of the equation).
one tau_n quantity can be handled, however, as soon as two (tau_n and tau_p)
quantities are mentioned in the denominator, the assembly becomes messed up.
-> KR