(quicksafe) working on generic and dynamic initial guesses
work in progress. each PDE set stores a map<quantity_name, initial_guess*>
the discretization method access retrieves this generic initial guess
for a particular unknown quantity (for nonlinear problems only ..)
and forwards the pointer to the underlying functor to viennafvm's set_initial_value method.

not entirely convinced with current implementation: virtual functor is called for each
cell, so for each cell a virtual function is called ..
also, the functor needs access to the device and the segment index it should operate on
(to access other possibly required quantities, e.g., for builtin potential we need the doping)
-> this feels clumsy. need to rethink