added FindUDUNITS.cmake script and corresponding flag
the udunits library is available in, for instance, Linux Mint 16.
So, this cmake module can find it and set the includes/libs accordingly.

however, the implementation of ViennaMini's quantity mechanism,
should also 'work' without the availability of udunits, albeit no automatic
unit handling but instead using the quantity 'as is'.