unit support implemented: automatic checks + conversions
using the udunits library. If this library is available on the system,
ViennaMini will not just check for non-conform units but will also try
to convert the units.

e.g. 1/cm3   will be automatically converted to  1/m3
however
     1 s cannot be converted to 1 K, as the units are not compatible.

This conversion mechanism is completly dealt with by udunits.

If the udunits library is not available on the system,
the automatic conversion mechanism is deactivated.
However, if a non-conform unit is detected, an exception is thrown asking the
user to correct the unit otherwise an actual simulation won't be conducted.

all examples work.