switched examples off by default and linked with ViennaMesh
as we are now switching towards the xml-driven application,
building the examples (which provide hard-coded device examples)
is not the primary use case anymore -> with this commit
by default, the examples are not generated.

Also, as we are using ViennaMesh currently only for one example,
the generation of the ViennaMesh library is linked to the condition
whether the examples are build -> this drastically reduces the build time
of ViennaMini.